+++ src/Autocomplete/Autocomplete.stories.tsx
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024 HCL America Inc.                                          *
+ * Copyright 2024, 2025 HCL America Inc.                                    *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -51,6 +51,15 @@ export default {
       },
       description:
         'Tooltip text hovering on ? mark for Autocomplete component',
+    },
+    enableHelpHoverEffect: {
+      control: 'boolean',
+      table: {
+        defaultValue: {
+          summary: false,
+        },
+      },
+      description: 'If true, the helper icon displays a gray background when hovered.',
     },
     tooltipPlacement: {
       description: 'Tooltip placement for ? mark for Autocomplete component.',
@@ -254,6 +263,7 @@ export const ExampleAutocomplete = {
     autoFocus: false,
     hiddenLabel: false,
     nonEdit: false,
+    enableHelpHoverEffect: false,
     actionProps: [
       // In component, this will render only max 2 action links at least until Figma design for action link variants is finalized
       {